pyproject.toml
# Copyright (c) 2026, Kashyapsinh Gohil
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-mech_interp"
version = "0.1.0"
description = "Mech Interp environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    # Core OpenEnv runtime (provides FastAPI server + HTTP client types)
    "openenv-core[core]>=0.2.2",
    # Web framework and server
    "fastapi>=0.115.0",
    "uvicorn>=0.24.0",
    # Environment-specific dependencies
    "torch>=2.0.0",
    "pydantic>=2.0.0",
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
# Server entry point - enables running via: python -m mech_interp.server.app
server = "mech_interp.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["mech_interp", "mech_interp.server"]
package-dir = { "mech_interp" = ".", "mech_interp.server" = "server" }